mikhailtal-manifold-agent-main/requirements.txt
requests
python-dotenv
numpy
pydantic
